# expression: implement vectorized evaluation for `builtinTimestampDiffSig`