<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

PCP #12103 

### What problem does this PR solve? <!--add issue link with summary if exists-->

implement vectorized evaluation for `builtinTimestampDiffSig`

### What is changed and how it works?

```
$ make vectorized-bench VB_FILE=Time VB_FUNC=builtinTimestampDiffSig
cd ./expression && \
                go test -v -benchmem \
                        -bench=BenchmarkVectorizedBuiltinTimeFunc \
                        -run=BenchmarkVectorizedBuiltinTimeFunc \
                        -args "builtinTimestampDiffSig"
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-8           1000000000               0.0192 ns/op          0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestampDiffSig-VecBuiltinFunc-8                32610             36193 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestampDiffSig-NonVecBuiltinFunc-8             15638             76852 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      3.712s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test